refactor: ExpenseViewService로 소비 조회 관련 분리
ExpenseConverter 함수 분리하여 리팩토링
수정된 서비스 코드에 맞게 ExpenseViewServiceTest 코드 수정 필요
Related to: #3
